rm resources tab; move file; fix logs and ms typos and font weights on faq
