+++ frontend/src/App.tsx
@@ -28,7 +28,6 @@ import {
   EXPLORE_DATA_PAGE_LINK,
   FAQ_TAB_LINK,
   OURTEAM_TAB_LINK,
-  RESOURCES_TAB_LINK,
   TERMS_OF_USE_PAGE_LINK,
   WHAT_IS_HEALTH_EQUITY_PAGE_LINK,
   NEWS_PAGE_LINK,
@@ -152,10 +151,6 @@ export default function App() {
                         </Route>
 
                         <Route path={FAQ_TAB_LINK}>
-                          <WhatIsHealthEquityPage />
-                        </Route>
-
-                        <Route path={RESOURCES_TAB_LINK}>
                           <WhatIsHealthEquityPage />
                         </Route>
