+++ frontend/src/pages/Methodology/methodologyContent/ResourcesData.ts
