+++ frontend/src/pages/Methodology/methodologySections/PdohLink.tsx
@@ -4,7 +4,7 @@ import {
   pdohDefinitionsArray,
 } from '../methodologyContent/PdohDefinitions'
 import Resources from '../methodologyComponents/Resources'
-import { PDOH_RESOURCES } from '../../WhatIsHealthEquity/ResourcesData'
+import { PDOH_RESOURCES } from '../methodologyContent/ResourcesData'
 import { Helmet } from 'react-helmet-async'
 import StripedTable from '../methodologyComponents/StripedTable'
 import { DATA_CATALOG_PAGE_LINK } from '../../../utils/internalRoutes'
