+++ frontend/src/utils/internalRoutes.ts
@@ -7,7 +7,6 @@ export const ABOUT_US_PAGE_LINK = '/aboutus'
 export const WHAT_IS_HEALTH_EQUITY_PAGE_LINK = '/whatishealthequity'
 export const TERMS_OF_USE_PAGE_LINK = '/termsofuse'
 export const FAQ_TAB_LINK = '/faqs'
-export const RESOURCES_TAB_LINK = '/resources'
 export const CONTACT_TAB_LINK = '/contact'
 export const OURTEAM_TAB_LINK = '/ourteam'
 export const SHARE_YOUR_STORY_TAB_LINK = '/shareyourstory'
